move memory.py from server/ to shared/
